# internal/controller/api/api.spec.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Cloud-Connector API",
      "url": "/api/cloud-connector/v1/"
    }
  ],
  "info": {
    "description": "Cloud-Connector Service",
    "version": "1.0.0",
    "title": "Cloud-Connector Service"
  },
...
  "paths": {
    "/message": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Submit a message to be routed to a customers environment.  This method will take the permitted accounts into consideration when looking up the connection details.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "PSKAuthClientID": [],
            "PSKAuthAccount": [],
            "PSKAuthKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "No connection to the target connected client"
          }
        }
      }
    },
    "/connection_status": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Retrieve the status of the connection.  This method will take the permitted accounts into consideration when looking up the connection details.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "PSKAuthClientID": [],
            "PSKAuthAccount": [],
            "PSKAuthKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "No connection to the target connected client"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-identity",
        "description": "Base64-encoded JSON identity header provided by 3Scale. Contains an account number of the user issuing the request. Format of the JSON: {\"identity\": {\"account_number\": \"12345678\"}}"
      },
      "PSKAuthClientID": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-client-id",
        "description": "Name of the service that is making the request"
      },
      "PSKAuthKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-psk",
        "description": "Key supplied by the platform infrastructure team"
      },
      "PSKAuthAccount": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-account",
        "description": "Account the request is being made on behalf of"
      }
    },
    "schemas": {
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MessageRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "recipient": {
            "type": "string"
          },
          "payload": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "directive": {
            "type": "string"
          }
        }
      },
...
      "ConnectionStatusRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          }
        }
      },
      "ConnectionStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "dispatchers": {
            "type": "object"
          }
        }
      },
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "connected",
          "disconnected"
        ]
      }
    }
  }
}